Mock/API/ServiceNow/requirements.txt
bcrypt==3.2.0
fastapi==0.111.0
passlib==1.7.4
PyJWT==2.8.0
uvicorn==0.30.1
requests==2.32.3
